tests/node-traversal.rs: added `test_node_prev_sibling`
